--- a/SelectQuery.Benchmarks/Program.cs
+++ b/SelectQuery.Benchmarks/Program.cs
@@ -1,0 +1,4 @@
+﻿using BenchmarkDotNet.Running;
+using SelectQuery.Benchmarks;
+
+var summary = BenchmarkRunner.Run<EvaluationBenchmarks>();