--- a/SelectQuery.Benchmarks/SelectQuery.Benchmarks.csproj
+++ b/SelectQuery.Benchmarks/SelectQuery.Benchmarks.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net8.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <PackageReference Include="BenchmarkDotNet" Version="0.14.0" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\SelectParser\SelectParser.csproj" />
+      <ProjectReference Include="..\SelectQuery.Evaluation\SelectQuery.Evaluation.csproj" />
+    </ItemGroup>
+
+</Project>