"large" dataset/"large IN filter" benchmark:
old - 709ms + 8,531MB memory allocated
new - 94ms + 231MB memory allocated


before:
![image](https://github.com/user-attachments/assets/ff3237e8-e69b-4075-a028-3b9aef38e290)


after:

| Method | Input | Query            | Mean          | Error         | StdDev       | Gen0        | Gen1       | Gen2      | Allocated     |
|------- |------ |----------------- |--------------:|--------------:|-------------:|------------:|-----------:|----------:|--------------:|
| **Run**    | **large** | **aggregated**       |  **90,362.94 μs** | **16,037.957 μs** |   **879.095 μs** |  **13500.0000** |  **2666.6667** |         **-** |  **248482.73 KB** |
| **Run**    | **large** | **filtered**         |  **88,598.45 μs** | **10,419.133 μs** |   **571.108 μs** |  **15333.3333** |  **2333.3333** |         **-** |  **281996.25 KB** |
| **Run**    | **large** | **large IN filter**  |  **94,312.09 μs** |  **1,729.001 μs** |    **94.772 μs** |  **12500.0000** |  **2333.3333** |         **-** |  **236278.06 KB** |
| **Run**    | **large** | **large projection** | **440,186.40 μs** | **54,968.911 μs** | **3,013.032 μs** | **183000.0000** | **21000.0000** |         **-** | **3459079.17 KB** |
| **Run**    | **large** | **projection**       |  **89,884.50 μs** | **13,150.913 μs** |   **720.846 μs** |  **13500.0000** |  **2500.0000** |         **-** |  **249459.45 KB** |
| **Run**    | **large** | **simple**           | **156,363.57 μs** | **11,670.918 μs** |   **639.722 μs** |  **12750.0000** |  **3000.0000** | **1000.0000** |  **306824.34 KB** |
| **Run**    | **small** | **aggregated**       |     **115.91 μs** |      **5.666 μs** |     **0.311 μs** |     **43.9453** |     **0.2441** |         **-** |     **807.66 KB** |
| **Run**    | **small** | **filtered**         |     **165.36 μs** |      **8.308 μs** |     **0.455 μs** |     **80.0781** |     **0.4883** |         **-** |    **1475.93 KB** |
| **Run**    | **small** | **large IN filter**  |      **80.36 μs** |      **3.645 μs** |     **0.200 μs** |     **25.5127** |     **0.1221** |         **-** |      **468.9 KB** |
| **Run**    | **small** | **large projection** |   **7,034.38 μs** |    **137.729 μs** |     **7.549 μs** |   **3703.1250** |    **23.4375** |         **-** |   **68122.51 KB** |
| **Run**    | **small** | **projection**       |     **136.17 μs** |     **11.358 μs** |     **0.623 μs** |     **44.6777** |     **0.4883** |         **-** |     **824.09 KB** |
| **Run**    | **small** | **simple**           |     **104.89 μs** |      **9.854 μs** |     **0.540 μs** |      **9.5215** |     **0.1221** |         **-** |     **175.08 KB** |